% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_dimensions.R
\name{device_dimensions}
\alias{device_dimensions}
\title{Get requested output dimensions}
\usage{
device_dimensions(default_text, height, aspect_ratio)
}
\arguments{
\item{default_text}{Default text size in points (1/72 of an inch)}

\item{height}{The desired height of the graphic, in lines of text}

\item{aspect_ratio}{The desired aspect ratio of the output graphic, as
\code{width / height}}
}
\value{
Two-element list containing the height and width in inches
}
\description{
Get the dimensions to pass to a graphics device (or related, such as
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}) for a graphic. The output proportions are
determined by the default text size, the height in lines of text, and the
aspect ratio.
}
\examples{
# A graphic at 16:9 and 35 lines of text tall
device_dimensions(
  default_text = 11,
  height = 35,
  aspect_ratio = 16/9
)
}
