% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ggunit_save}
\alias{ggunit_save}
\title{Save a file quickly with ggunit}
\usage{
ggunit_save(..., default_text = 11, height, aspect_ratio)
}
\arguments{
\item{...}{Passed to ggsave. Note that \code{height} and \code{width} are generated
by this function.}

\item{default_text}{default text size}

\item{height}{Height of the graphic in lines}

\item{aspect_ratio}{Aspect ratio in the form width/height}
}
\description{
This function is called for its side-effects. It wraps \code{ggplot2::ggsave()}
to allow you save without first running \code{device_dimensions()}.
}
