% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{mm}
\alias{mm}
\title{Convert any units to millimeters}
\usage{
mm(x, units = "pts")
}
\arguments{
\item{x}{A numeric vector}

\item{units}{A character vector specifying the units for the corresponding
numeric values.}
}
\value{
The same number converted to millimeters
}
\description{
Convert any units to millimeters
}
