% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{pt_to_mm}
\alias{pt_to_mm}
\alias{mm_to_pt}
\title{Convert points and millimeters}
\usage{
pt_to_mm(x)

mm_to_pt(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} uses millimeters to measure text size, but
\code{\link[ggplot2:element]{ggplot2::element_text()}} uses
points. Hence, you may wish to convert to one or the other to maintain
standard sizes.
}
