% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{px}
\alias{px}
\title{Convert from any units to pixels}
\usage{
px(x, units = "pts", res = 96, round = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{units}{A character vector specifying the units for the corresponding
numeric values.}

\item{res}{Pixels per inch to use as resulution. Defaults to 96, because this
is a common web font conversion factor.}

\item{round}{Whether to round to the nearest full pixel. Defaults to \code{TRUE}.}
}
\description{
Convert from any units to pixels
}
